pyproject.toml
[project]
name = "dynamic-chart-builder"
version = "0.1.0"
description = "Dynamic Chart Builder with Django DRF"
authors = [
    {name = "Your Name",email = "your.email@example.com"}
]
readme = "README.md"
requires-python = "^3.10"
package-mode = false
dependencies = [
    "django (>=5.2.7,<6.0.0)",
    "djangorestframework (>=3.16.1,<4.0.0)",
    "django-cors-headers (>=4.9.0,<5.0.0)",
    "psycopg2-binary (>=2.9.11,<3.0.0)",
    "python-dotenv (>=1.1.1,<2.0.0)",
    "gunicorn (>=23.0.0,<24.0.0)",
    "drf-spectacular (>=0.28.0,<0.29.0)",
    "django-filter (>=25.2,<26.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[dependency-groups]
dev = [
    "pytest (>=8.4.2,<9.0.0)",
    "pytest-django (>=4.11.1,<5.0.0)",
    "black (>=25.9.0,<26.0.0)",
    "flake8 (>=7.3.0,<8.0.0)",
    "pytest-cov (>=7.0.0,<8.0.0)"
]
